refactor(ingestion): convert ingest_ipdb to plan/apply framework
Splits the 1,250-line ingest_ipdb management command into:
- ipdb/adapter.py: build_ipdb_plan() — non-mutating adapter that
  produces an IngestPlan for CorporateEntity, MachineModel, Person,
  and Theme entities
- ipdb/features.py: feature/theme/reward extraction (pure parsing)
- ingest_ipdb.py: thin wrapper (~70 lines) calling build → apply

Deliberate behavior changes:
- Sweep dropped (additive-only, matching OPDB conversion)
- Dead manufacturer claim on MachineModel replaced with
  ipdb.corporate_entity_name and ipdb.manufacturer_trade_name
  extra_data claims
- Manufacturers must pre-exist (no auto-create)
- Location validation skipped for newly-created CEs

Uses identity_refs for deferred credit/theme relationship claims
where Person/Theme entities are created in the same plan.
